Added userspace device mapper library
This adds a userspace library to talk to the Kernel's DM subsystem and
is part of the Smalltown init to set up dm-integrity/dm-crypt.

Test Plan:
Currently manually tested, automated testing possible but would require
spinning up a kernel and testing against it. This would require KVM access
on the test infrastructure, a test kernel and additional code.

X-Origin-Diff: phab/D154
GitOrigin-RevId: 45565ae6288e2accee3f8ce80233580c6ac3e754