Delete old secretstore, cmd/node and config modules
This removes even more code that is no longer necessary or interferes with new concepts. It also refactors the storage stuff into a StorageManager which deals with all the paths and async initialization.

This does intentionally break a few things which will be fixed when the CA code lands.

Test Plan: Manually tested using make launch, CI is in a separate ticket

X-Origin-Diff: phab/D182
GitOrigin-RevId: 282a4bd84b47010d859e03da53b2c2de8183b13b