Increase robustness of bazel wrapper script and remove repository_cache
This allows the wrapper script to be called from anywhere.

The repo cache does not actually work as expected since it does not
cache most external dependencies we care about, and is not always a
valid command line argument, so we would either have to specify it
in .bazelrc (breaking non-standard dev setups), or specify it manually.

Test Plan: Ran the bazel wrapper from my home directory, got expected output.

X-Origin-Diff: phab/D204
GitOrigin-RevId: 74d09ba24fd84ba0dd6e1ba282995c452546eb25