Separate cache volume for each working copy
We can rely on the repository leases to coordinate Bazel caches,
basically leasing them out along with the repository.

With multiple copies of the cache, keeping them on a tmpfs seems
like a bad idea, so we switch to a local volume.

Test Plan: Covered by CI

X-Origin-Diff: phab/D250
GitOrigin-RevId: 01d9392198d8c00089d3133425091ab766b9b590