Added kube-controlplane binary
This adds a custom binary which contains all Kubernetes control plane
components. This is necessary since every control plane binary by itself
is around 130MiB and this combined one is only around 150MiB. This
can be cut in half to around 70MiB as soon as Kubernetes can be built
providerless by Bazel.

I'm not entirely happy with the integration, we may need gazelle
exclusions and a plan to deal with go mod since it can't resolve the
dependencies in a reasonable way.

Test Plan: Manual test with kubectl (this by itself is not runnable)

Bug: T485

X-Origin-Diff: phab/D256
GitOrigin-RevId: d76702f2cd0d71463ff891e5a44eac7b66be07f0