Remove synthetic GOPATH and symlinks
As it turns out, `go mod tidy` complains about missing modules
but still properly generates go.mod, so it appears that we do
not need a GOPATH or symlinks except for non-IntelliJ IDE support.

We should standardize on IntelliJ and just get rid of this
workaround. Hendrik had trouble rebasing the shield revisions
(that have a lot of generated code) and it's likely a lot of
work to maintain this hack in the future.

See D305, D309

Test Plan:
Ran `scripts/gazelle.sh` after deleting all symlinks
and tested that `go mod tidy` properly restored the go.mod file
after modifying it manually.

X-Origin-Diff: phab/D311
GitOrigin-RevId: fe4bc491fca6fa072cff047185d7c18305564ea4