core/internal/api: use gRPC statuses as much as possible
Returning plain go errors via gRPC will always result in a gRPC 'INTERNAL' error code, which is suboptimal. We go ahead and semanticize some of the possible error paths, and at the same time:

 - swallow some internal errors into logs and serve sanitized errors
 - move some of the internal service implementations to also use gRPC statuses
 - change a panic() call into a status.Unimplemented return type

There's still plenty work to be done on this front, but this is a good first change.

Test Plan: if this is not covered by tests we're screwed anyways

X-Origin-Diff: phab/D366
GitOrigin-RevId: 71880a9e23c65631d6c4df6338855864c34bb11f