third_party: slurp in edk2, kubernetes, mkfs.xfs
This finishes the move from core/build/* into third_party/.

While at first this might look like wasted bandwidth, this separation
will make much more sense in the future, where different parts (not only
the Smalltown core) might depend on shared external dependencies. In
addition, having everything in third_party laid out in a similar fashion
lends itself to writing more general rules. Already there is quite a bit
of deduplicaiton that we could remove for reliability and readability.

This does not fix the problem of the big honkin' genrule for mkfs.xfs -
while I think we should fix it sooner than later by building a real
toolchain, that time is not yet now. But at least we've moved things out
of the way so that we can then drop in a better mkfs.xfs, once it is
built so.

Test Plan: build file mangling, CI should cover this

X-Origin-Diff: phab/D391
GitOrigin-RevId: fb99c6a6270c5c6a56eeb4f18a41323ffebbc655