supervisor: init
This introduces the service supervisor (or supervisor, for short) - a
library used to reliably run parts of Smalltown.

The design is outlined in [[ https://phab.monogon.dev/u/supervision | go/supervision ]].

This only implements the supervision itself, and does not actually use
it in Smalltown. Another revision based on this one will aims to move at
least parts of the codebase onto this library.

Test Plan: the supervision code is integration tested

Bug: T653

X-Origin-Diff: phab/D429
GitOrigin-RevId: cffa73de5957e95af629b78379ffc0c7e8681afb