core/internal/common/supervisor: deflake
We fix several flaky tests in the supervisor framework, and one bug in the
supervisor itself.

Tests are deflaked by depending less on tight timing and instead on a
'settled' state of the supervisor, which is basically a wait to join the
supervisor when it's done with whatever work it's currently taking care
of.

Another flake, TestBackoff, is fixed by widening the allowed restart
time.

Finally, we fix a bug in the supervisor that caused it to spuriously
restart children when it would schedule them when their future parents
context was canceled.

Finally, we make some log messages less verbose.

Test Plan: Covered by existing tests that are now less flaky. This was tested with bazel test --runs_per_test=100 to not flake anymore.

X-Origin-Diff: phab/D495
GitOrigin-RevId: f92f7368708c54c59644d3e7dca03b2b5692c30a