*: unpin rules_go/gazelle
We unpin these now that Gazelle 0.21.0 is out. This release also changes
runfile pathnames for go_binary, so we have to fix up the test_boot.sh
script that hardcodes a path.

This also pulls in a new protobuf version, that we have to let in with
its imperfections that nogo rightfully complains about.

Test Plan: machinery change, current tests cover this

X-Origin-Diff: phab/D538
GitOrigin-RevId: ad83d7868608b6883a891d127a6fbaf28f8aa14a