build/toolchain: init
This adds a new, bare-bones, host-based C++ toolchain, and enables it
for all target builds. This toolchain replaces the automatically
generated host toolchain from bazel-tools, and differs in the following
ways:
 - uses fully hardcoded paths
 - is the bare minimum required, which allows us full control over all
   aspects of it, notably link arguments
 - does not assume we're building normal C++ binaries for Linux (for
   instance, the new toolchain does not always link with -lm)

This is in anticipation of a change by @lorenz that uses cc_binary to
build qboot for tests. However, this is also a good basis to start
writing a 'real' toolchain suite for mkfs.xfs, linux & co.

Test Plan: For now, this is unused - but does not break any existing flow, which is fine. I did test this on a qboot WIP commit from @lorenz, and it at least fixed our immediate problem that it wanted to build it with with -lm,-lstdc++ - which we didn't.

X-Origin-Diff: phab/D560
GitOrigin-RevId: 8a5bc5f00a0a0534ea245e556d160f5bab7f8a0c