scripts/create_container: fix cockroachdb startup
After moving the build container to --net=host this broke building //...
(as sqlboiler touches a local crdb in order to generate SQL
boilerplate...). This moves cockroachdb to also run with --net=host, and
fixes the advertisement address in the same way as it's fixed in
run_ci.sh.

Test Plan: tested this locally :/

X-Origin-Diff: phab/D562
GitOrigin-RevId: 8b4dcc2b5c3f87c48824d6e0eb665bea44e84116