core/internal/localstorage: init
This implements localstorage and localstorage/declarative, a small
library for better typed filesystem access. Further down the road this
will replace //core/internal/storage, but we're trying to commit this
early.

This is not used anywhere, and instead comes with a basic test to show
its workings.

Test Plan: covered by unit tests

X-Origin-Diff: phab/D578
GitOrigin-RevId: 9a225bc105cc331ce139eb6c195e9af216c8633e