Kubernetes volume & stamping fixes
This reenables the projected and downwardapi volume types, both of which are necessary for the CTS to pass
and are derivatives of the configmap volume which we already support. This also fixes an issue where the stamping
definitions for Kubernetes were not present on our main Kubernetes binary, which broke the CTS.

Test Plan:
Volumes will be covered by CTS (writing our own tests would be duplicate work), version was manually
tested to be correct in `bazel run //core/cmd/dbg -- kubectl get nodes` since stamping is hard to test for.

X-Origin-Diff: phab/D584
GitOrigin-RevId: 403b6c845bc399fdd44ec3ba4ca26e2512a5bc98