core/api: move to core/proto
This is keeping in line with conventions that protobuf files generally
live in a 'proto/' directory. Even without that, a lot of the protos in
there aren't actually part of an API, so keeping them in `api/` is a bit
of a misnomer.

We also remove unused protos that were part of the old
integrity/lifecycle flow. Again, these will make a comeback.

Test Plan: this should fail. part of a larger stack. D590 is the first tip of the stack that should work.

X-Origin-Diff: phab/D588
GitOrigin-RevId: 4a7af494810348f6bcabd49e63902b4c47e6ec35