cluster/internal: remove old cluster enrolment/integrity/management
This had to be done sooner or later, as it has been woefully
underdesigned. A lot of the TPM2 code will make a comeback, but keeping
this around (and buildable) right now is too painful. Once we get
multi-node clusters again, and properly design node/cluster lifecycle,
we'll add integrity/attestation support back in.

Test Plan: this should fail. part of a larger stack. D590 is the first tip of the stack that should work.

X-Origin-Diff: phab/D587
GitOrigin-RevId: e8a43906a767aa4cb66b051027d619ce364269e7