supervisor: never give up
This fixes T756, in which supervised processes would reach a negative
backoff value. This seems to be caused by the backoff library's
ExponentialBackoff having a default MaxElapsedTime of 15 minutes, after
which it returns 'Stop', or, -1 seconds.

Test Plan: There's no easy way to test this. Unfortunately, the behaviour to return Stop is not after a number of calls, but after time has elapsed. We don't want to wait 15 minutes for a test, and we don't have an easy way to mock time, either. But I did test this manually and I cannot observe the 'negative backoffs' after 15 minutes anymore.

Bug: T756

X-Origin-Diff: phab/D619
GitOrigin-RevId: 49d8617bcf2c8b36127cb43acde8afb7cc35c99f