Fix nanoswitch not setting DHCP lease time
The original implementation of this used an accessor (due to
confusing naming on dhcpv4's part) and passed the value to be set the
default value. Found when testing with the new DHCP client which
correctly rejected these leases as invalid.

Test Plan: Works with the new DHCP client and in E2E

X-Origin-Diff: phab/D654
GitOrigin-RevId: 841b161a02809d3c43768ff7a529905f90290f2c