Make LaunchCluster not close DebugClient it's returning
LaunchCluster currently closes the gRPC connection underpinning the DebugService
it's returning. This causes all further calls to that service to immediately return with
a CANCELED error. There's no reason why it should do this (probably a refactoring artifact),
so just remove the Close call. Fixes T881.

Test Plan: CTS is not run as part of automated test still because it takes too long.

Bug: T881

X-Origin-Diff: phab/D671
GitOrigin-RevId: 6bfa382cba6a15b146b2f24311507456b58cdf98