Drop legacy kubelet log path
It looks like the Kubernetes update broke E2E tests on the EROFS stack because of some change
to how the legacy log dir is handled. Kubelet currently just crashes because it can't mkdir
/var/log/containers. This directory is apparently only used by fluentd for log collection in upstream
E2E tests and with dockershim, both of which we don't care about. So let's just nuke it.

Test Plan: Fixes things on top of the EROFS stack

X-Origin-Diff: phab/D700
GitOrigin-RevId: 45b7f76a61b7234845ab55fcfbc37a66f69fe065