m/node/core/conensus: handle etcd restarts
This makes the etcd service more reliable in case of transient failures
when starting in a new cluster. Previously, any restart of etcd on the
first (bootstrapping) node would cause etcd to get stuck and never start
again (as certificates were already created). This changes the logic to
allow existing certificates.

This also handles the case of etcd attempting to start as the network is
reconfigured, and eg. the external hostname is not yet resolvable.

Test Plan:
No tests yet. This should be tested by a more comprehensive e2e test where we
randomly kill some runnables (see: T872).

X-Origin-Diff: phab/D733
GitOrigin-RevId: 8ac426f9423ec2353537eec651071e99a5e5ec53