Include compiler-specific headers in musl toolchain sysroot
Currently our musl sysroot does not include any compiler-specific headers. This makes the musl
toolchain unable to build more complicated things like QEMU. Since the musl toolchain is not actually
a real toolchain, but just a thin wrapper around the host toolchain this is a problem as these headers
are not static. For a lack of better options it has thus been decided that we're just going to manually
merge the headers for our build container in.

This is expected to be removed as soon as we have a proper toolchain.

Test Plan: Tested in subsequent revision

X-Origin-Diff: phab/D737
GitOrigin-RevId: a1ed1d229c87980341c80b22263f5d9a4cf9924c