third_party/qemu: better explain tb_invalidate_phys_range patch
At a glance, this change looks somewhat scary, and might be interpreted
as an attempt to backdoor qemu.

This better explains what's going on, and adds an extra always-firing
assert to prove that there's nothing up our sleeves, and that this
branch should never be taken in the first place.

Test Plan: Refactor, should be covered by tests.

X-Origin-Diff: phab/D744
GitOrigin-RevId: c86638cf9e90041d2ad19d26715c7d4dd5a43e98